+++ src/negmas/helpers/__init__.py
@@ -15,6 +15,7 @@ from .timeout import *
 from .misc import *
 from .numba_checks import *
 from .parallel import *
+from .rand import *
 from . import parallel
 
 __all__ = (
@@ -25,6 +26,7 @@ __all__ = (
     + misc.__all__
     + numba_checks.__all__
     + parallel.__all__
+    + rand.__all__
     # + ["prob", "inout", "numeric"]
     # + numeric.__all__
     # + prob.__all__